requirements.txt
Django==4.2
django-cors-headers==3.14.0
django-environ==0.9.0
djangorestframework==3.14.0
docker==5.0.3
docker-compose==1.29.2